Refactor common JSON methods to new helper namespace
